Extract method to a separate helper
Since it's not directly related to participatory processes.
