+++ decidim-core/app/controllers/decidim/participatory_process_steps_controller.rb
@@ -11,6 +11,7 @@ module Decidim
     include NeedsParticipatoryProcess
 
     helper Decidim::ParticipatoryProcessHelper
+    helper Decidim::IconHelper
 
     def index
       authorize! :read, ParticipatoryProcess
