+++ decidim-core/app/helpers/decidim/participatory_process_helper.rb
@@ -13,19 +13,5 @@ module Decidim
       dates = [participatory_process_step.start_date, participatory_process_step.end_date]
       dates.map { |date| date ? localize(date.to_date, format: :default) : "?" }.join(" - ")
     end
-
-    # Public: Returns an icon given an instance of a Feature. It defaults to
-    # a question mark when no icon is found.
-    #
-    # feature - The feature to generate the icon for.
-    #
-    # Returns an HTML tag with the icon.
-    def feature_icon(feature)
-      if feature.manifest.icon
-        external_icon feature.manifest.icon
-      else
-        icon "question-mark"
-      end
-    end
   end
 end
